Update show identity check retry condition
Small change that will allow to not break the app on new version of Moonpay API. Let me know when it’s merged and deployed in a new RC.

Explanation: `rejectType` has now three possible values.

1. `final`: Final reject
2. `retry`: A reject that can be fixed by uploading a new image
3. `external`: Everything is fine with submitted documents but there are some external reasons to reject, such as address not matching or date of birth not matching

It’s allowed to update Customer object and upload files for the reasons 2. and 3. but you can improve the flow in the future, only asking the user for the rejected informations.